requirements.txt
django==2.2.1
psd-tools==1.8.15
kombu==4.5.0
python-decouple==3.1
djmail==1.1.0
django-jinja==2.4.1
django-sr==0.0.4
djangorestframework==3.9.4
easy-thumbnails==2.6
django-storages==1.7.1
django-extensions==2.1.7
psycopg2==2.8.2
django-money==0.14.4
django-pglocks==1.0.2
Unidecode==1.0.23
premailer==3.4.1
django-ipware==2.1.0
python-oauth2==1.1.0
django-sites==0.10
qrcode==6.1
gunicorn==19.9.0
pygeoip==0.3.2
boto3==1.9.156
blockchain==1.4.4
django-celery-results==1.1.2
django-celery-beat==1.5.0
django-debug-toolbar
django-bootstrap-form
django-user-sessions
ecdsa
requests
django-picklefield
django-countries
rest-pandas
xlwt
